requirements.txt
django==5.0.3
psycopg2-binary==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
load_dotenv==0.1.0
pillow==10.2.0
django-cors-headers==4.3.1
pyTelegramBotAPI==4.16.1
TELEGRAM_TOKEN=6258725371:AAFYekZZyzGEUO8jOUtSQFa07JnGQBdOCQw
celery==5.3.6
eventlet==0.36.0
redis==5.0.3
django-celery-beat==2.6.0